Add aliasing!
Aliases are recursive, deleting a gif that's an alias deletes the alias but
lets you know what it was pointing to in case you messed up, deleting an alias
that's actually a gif container errors out, adding a gif to an alias lets you
know what it points to so you can update the original, trying to remove a
specific URL under an alias likewise lets you know to update the original.

It's a ball.